requirements.txt
# Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
python-multipart==0.0.6

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Frontend
streamlit==1.40.1
plotly==5.24.1
pandas==2.2.3

# Database
alembic==1.14.0

# Utilities
python-dateutil==2.9.0
requests==2.32.3
openpyxl==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
